# Conflicts:
#	lib/app/base_controller.dart
#	lib/app/modules/chat/controllers/chat_controller.dart

+++ lib/app/modules/group/controllers/group_info_controller.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:mirror_fly_demo/app/common/constants.dart';
+import 'package:mirror_fly_demo/app/common/main_controller.dart';
 import 'package:mirror_fly_demo/app/data/helper.dart';
 import 'package:mirrorfly_plugin/mirrorfly.dart';
 
@@ -15,6 +16,7 @@ import '../../../routes/app_pages.dart';
 import '../views/name_change_view.dart';
 
 class GroupInfoController extends GetxController {
+  var availableFeatures = Get.find<MainController>().availableFeature;
   ScrollController scrollController = ScrollController();
   var groupMembers = <Profile>[].obs;
   final _mute = false.obs;
@@ -26,8 +28,8 @@ class GroupInfoController extends GetxController {
   bool get isAdmin => _isAdmin.value;
 
   final _isMemberOfGroup = true.obs;
-  set isMemberOfGroup(value) => _isMemberOfGroup.value=value;
-  bool get isMemberOfGroup => _isMemberOfGroup.value;
+  set isMemberOfGroup(value) => _isMemberOfGroup.value = value;
+  bool get isMemberOfGroup => availableFeatures.value.isGroupChatAvailable.checkNull() && _isMemberOfGroup.value;
 
   var profile_ = Profile().obs;
   //set profile(value) => _profile.value = value;
@@ -44,7 +46,9 @@ class GroupInfoController extends GetxController {
     _mute(profile.isMuted!);
     scrollController.addListener(_scrollListener);
     getGroupMembers(false);
-    getGroupMembers(null);
+    if(availableFeatures.value.isGroupChatAvailable.checkNull()){
+      getGroupMembers(null);
+    }
     groupAdmin();
     memberOfGroup();
     muteAble();
@@ -166,17 +170,21 @@ class GroupInfoController extends GetxController {
     });
   }
   memberOfGroup(){
-    Mirrorfly.isMemberOfGroup(profile.jid.checkNull(),null).then((bool? value){
+    Mirrorfly.isMemberOfGroup(profile.jid.checkNull(),SessionManagement.getUserJID()).then((bool? value){
       if(value!=null){
         _isMemberOfGroup(value);
       }
     });
   }
   onToggleChange(bool value){
-    if(muteable.value) {
-      mirrorFlyLog("change", value.toString());
-      _mute(value);
-      Mirrorfly.updateChatMuteStatus(profile.jid.checkNull(), value);
+    if (isMemberOfGroup) {
+      if (muteable.value) {
+        mirrorFlyLog("change", value.toString());
+        _mute(value);
+        Mirrorfly.updateChatMuteStatus(profile.jid.checkNull(), value);
+      }
+    }else{
+      toToast("You're no longer a participant in this group");
     }
   }
 
@@ -192,6 +200,10 @@ class GroupInfoController extends GetxController {
   }
 
   reportGroup(){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     Helper.showAlert(title: "Report this group?",message: "The last 5 messages from this group will be forwarded to admin. No one in this group will be notified.",actions: [
       TextButton(
           onPressed: () {
@@ -228,6 +240,10 @@ class GroupInfoController extends GetxController {
     }
   }
   leaveGroup(){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     Helper.showAlert(message: "Are you sure you want to leave from group?.",actions: [
       TextButton(
           onPressed: () {
@@ -242,7 +258,12 @@ class GroupInfoController extends GetxController {
           child: const Text("LEAVE")),
     ]);
   }
+  var leavedGroup = false.obs;
   exitFromGroup()async{
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(await AppUtils.isNetConnected()) {
       Helper.progressLoading();
       Mirrorfly.leaveFromGroup(SessionManagement.getUserJID() ,profile.jid.checkNull()).then((value) {
@@ -250,6 +271,7 @@ class GroupInfoController extends GetxController {
         if(value!=null){
           if(value){
             _isMemberOfGroup(!value);
+            leavedGroup(value);
           }
         }
       }).catchError((error) {
@@ -260,6 +282,10 @@ class GroupInfoController extends GetxController {
     }
   }
   deleteGroup(){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull() || !availableFeatures.value.isDeleteChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     Helper.showAlert(message: "Are you sure you want to delete this group?.",actions: [
       TextButton(
           onPressed: () {
@@ -270,6 +296,10 @@ class GroupInfoController extends GetxController {
           onPressed: () async {
             if(await AppUtils.isNetConnected()) {
               Get.back();
+              if(!availableFeatures.value.isGroupChatAvailable.checkNull() || !availableFeatures.value.isDeleteChatAvailable.checkNull()){
+                Helper.showFeatureUnavailable();
+                return;
+              }
               Helper.progressLoading();
               Mirrorfly.deleteGroup(profile.jid.checkNull()).then((value) {
                 Helper.hideLoading();
@@ -316,6 +346,10 @@ class GroupInfoController extends GetxController {
 
   final ImagePicker _picker = ImagePicker();
   camera() async {
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(await AppUtils.isNetConnected()) {
       final XFile? photo = await _picker.pickImage(
           source: ImageSource.camera);
@@ -339,6 +373,10 @@ class GroupInfoController extends GetxController {
   }
 
   updateGroupProfileImage(String path){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     showLoader();
     Mirrorfly.updateGroupProfileImage(profile.jid.checkNull(),path).then((bool? value){
       hideLoader();
@@ -352,6 +390,10 @@ class GroupInfoController extends GetxController {
   }
 
   updateGroupName(String name){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     showLoader();
     Mirrorfly.updateGroupName(profile.jid.checkNull(),name).then((bool? value){
       hideLoader();
@@ -382,6 +424,10 @@ class GroupInfoController extends GetxController {
   }
 
   revokeAccessForProfileImage()async{
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(await AppUtils.isNetConnected()) {
       showLoader();
       Mirrorfly.removeGroupProfileImage(profile.jid.checkNull()).then((bool? value) {
@@ -408,6 +454,10 @@ class GroupInfoController extends GetxController {
   }
 
   gotoAddParticipants(){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     Get.toNamed(Routes.contacts, arguments: {"forward" : false,"group":true,"groupJid":profile.jid })?.then((value){
       if(value!=null){
         addUsers(value);
@@ -436,6 +486,10 @@ class GroupInfoController extends GetxController {
   }
 
   removeUser(String userJid) async {
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(isMemberOfGroup){
       if(await AppUtils.isNetConnected()) {
         showLoader();
@@ -450,10 +504,16 @@ class GroupInfoController extends GetxController {
       }else{
         toToast(Constants.noInternetConnection);
       }
+    }else{
+      toToast("You're no longer a participant in this group");
     }
   }
 
   makeAdmin(String userJid) async {
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(isMemberOfGroup){
       if(await AppUtils.isNetConnected()) {
         showLoader();
@@ -468,11 +528,17 @@ class GroupInfoController extends GetxController {
       }else{
         toToast(Constants.noInternetConnection);
       }
+    }else{
+      toToast("You're no longer a participant in this group");
     }
   }
 
   //New Name Change
   gotoNameEdit(){
+    if(!availableFeatures.value.isGroupChatAvailable.checkNull()){
+      Helper.showFeatureUnavailable();
+      return;
+    }
     if(isMemberOfGroup) {
       Get.to(const NameChangeView())?.then((value) {
         if (value != null) {
@@ -506,5 +572,12 @@ class GroupInfoController extends GetxController {
 
   void userBlockedMe(String jid) {
     userUpdatedHisProfile(jid);
+  }
+
+  void onAvailableFeaturesUpdated(AvailableFeatures features) {
+    LogMessage.d("GroupInfo", "onAvailableFeaturesUpdated ${features.toJson()}");
+    availableFeatures(features);
+    _isMemberOfGroup.refresh();
+    // loadGroupExistence();
   }
 }
